+++ HISTORY.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Add new upstream MIME types
+
 1.52.0 / 2022-02-21
 ===================