+++ HISTORY.md
@@ -5,6 +5,7 @@ unreleased
   * Add extensions from IANA for more MIME types
   * Add new upstream MIME types
   * Fix extensions for `text/markdown` to match IANA
+  * Remove extension `.mjs` from `application/javascript`
   * Remove obsolete MIME types from IANA data
 
 1.52.0 / 2022-02-21