+++ HISTORY.md
@@ -4,6 +4,7 @@ unreleased
   * Add extensions `.aac` and `.adts` to `audio/aac`
   * Add extensions `.js` and `.mjs` to `text/javascript`
   * Add extensions from IANA for more MIME types
+  * Add Microsoft app installer types and extensions
   * Add new upstream MIME types
   * Fix extensions for `text/markdown` to match IANA
   * Remove extension `.mjs` from `application/javascript`