+++ HISTORY.md
@@ -50,7 +50,7 @@ unreleased
 ===================
 
   * Add new upstream MIME types
-  * Remove ambigious extensions from IANA for `application/*+xml` types
+  * Remove ambiguous extensions from IANA for `application/*+xml` types
   * Update primary extension to `.es` for `application/ecmascript`
 
 1.46.0 / 2021-02-13